Rename `midddleware` to `middleware` in examples.
