+++ examples/asynchronous_telebot/middleware/i18n_middleware_example/main.py
@@ -56,7 +56,7 @@ import keyboards
 from telebot import types
 from telebot.async_telebot import AsyncTeleBot
 from telebot.asyncio_filters import TextMatchFilter, TextFilter
-from i18n_base_midddleware import I18N
+from i18n_base_middleware import I18N
 from telebot.asyncio_storage.memory_storage import StateMemoryStorage
