+++ examples/middleware/class_based/i18n_middleware/i18n_base_middleware.py
