+++ examples/middleware/class_based/i18n_middleware/main.py
@@ -53,7 +53,7 @@ import asyncio
 from typing import Union
 
 import keyboards
-from i18n_base_midddleware import I18N
+from i18n_base_middleware import I18N
 from telebot import TeleBot
 from telebot import types, StateMemoryStorage
 from telebot.custom_filters import TextMatchFilter, TextFilter
